build: fix not found /packages/cardano-services/node_modules in nix flake

+++ flake.nix
@@ -69,7 +69,7 @@
 
           mkdir -p $out/libexec/$sourceRoot $out/bin
 
-          cp -r ${production-deps}/libexec/$sourceRoot/node_modules $out/libexec/$sourceRoot/node_modules
+          test -f ${production-deps}/libexec/$sourceRoot/packages/cardano-services/node_modules && cp -r ${production-deps}/libexec/$sourceRoot/node_modules $out/libexec/$sourceRoot/node_modules
           cp -r scripts $out/libexec/$sourceRoot/scripts
           for p in cardano-services core ogmios util; do
             mkdir -p $out/libexec/$sourceRoot/packages/$p
